+++ eras/shelley/impl/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Version history for `cardano-ledger-shelley`
 
-## 1.0.1.0
+## 1.1.0.0
 
-* Add `ToJSON` instance for `ShelleyTxOut`
-* Add `ToJSON` instance for `AlonzoPParams StrictMaybe`
+* Add `ToJSON` instance for:
+  * `ShelleyTxOut`
+  * `AlonzoPParams StrictMaybe`
+  * `ProposedPPUpdates` and `ShelleyPPUPState`
+* Add `ToJSON (GovernanceState era)` superclass constraint for `EraCovernance`
 
 ## 1.0.0.0